Split indexed pipeline policies
