+++ src/FAI.Core/Pipelines/OrderingPipeline.cs
@@ -1,0 +1,65 @@
+namespace FAI.Core.Pipelines;
+
+public sealed class OrderingPipeline<TInput, TOutput> : IPreallocatingPipeline<TInput, TOutput>
+{
+    private readonly IPipeline<TInput, TOutput> _inner;
+    private readonly IPreallocatingPipeline<TInput, TOutput>? _preallocatingInner;
+    private readonly IIndexOrdering<TInput> _ordering;
+    private readonly IReadOnlyIndexedBatch<TInput> _inputBatch;
+    private readonly IWritableIndexedBatch<TOutput> _outputBatch;
+
+    public OrderingPipeline(IPipeline<TInput, TOutput> inner, IIndexOrdering<TInput> ordering,
+        IReadOnlyIndexedBatch<TInput> inputBatch, IWritableIndexedBatch<TOutput> outputBatch)
+    {
+        _inner = inner;
+        _preallocatingInner = inner as IPreallocatingPipeline<TInput, TOutput>;
+        _ordering = ordering;
+        _inputBatch = inputBatch;
+        _outputBatch = outputBatch;
+    }
+
+    public bool TryAllocateOutput(TInput input, out TOutput output)
+    {
+        if (_preallocatingInner is not null && _preallocatingInner.TryAllocateOutput(input, out TOutput? allocated))
+        {
+            output = allocated;
+            return true;
+        }
+
+        output = default!;
+        return false;
+    }
+
+    public async ValueTask<TOutput> ExecuteAsync(TInput input, CancellationToken cancellationToken = default)
+    {
+        int[] sortedToOriginal = _ordering.CreateOrder(input);
+        using BatchLease<TInput> sortedInput = _inputBatch.Gather(input, sortedToOriginal);
+        TOutput output = await _inner.ExecuteAsync(sortedInput.Value, cancellationToken);
+        _outputBatch.PermuteInPlace(output, sortedToOriginal);
+        return output;
+    }
+
+    public async ValueTask ExecuteAsync(TInput input, TOutput output, CancellationToken cancellationToken = default)
+    {
+        int[] sortedToOriginal = _ordering.CreateOrder(input);
+        using BatchLease<TInput> sortedInput = _inputBatch.Gather(input, sortedToOriginal);
+        if (_preallocatingInner is not null)
+        {
+            await _preallocatingInner.ExecuteAsync(sortedInput.Value, output, cancellationToken);
+        }
+        else
+        {
+            TOutput sortedOutput = await _inner.ExecuteAsync(sortedInput.Value, cancellationToken);
+            try
+            {
+                int outputCount = _outputBatch.Count(sortedOutput);
+                Span<int> identity = stackalloc int[outputCount];
+                for (int index = 0; index < outputCount; index++) identity[index] = index;
+                _outputBatch.Scatter(sortedOutput, output, identity);
+            }
+            finally { await PipelineOutputDisposer.DisposeAsync(sortedOutput); }
+        }
+
+        _outputBatch.PermuteInPlace(output, sortedToOriginal);
+    }
+}
